+++ internal/pkg/tui/filter_operations.go
@@ -31,7 +31,7 @@ func (m Model) handleFilterInput(msg tea.KeyMsg) (tea.Model, tea.Cmd) {
 			// Empty filter = clear all filters
 			m.packetStore.ClearFilter()
 			m.packetStore.FilteredPackets = make([]components.PacketDisplay, 0)
-			m.packetStore.MatchedPackets = m.packetStore.PacketsCount
+			m.packetStore.MatchedPackets = int64(m.packetStore.PacketsCount)
 			m.uiState.PacketList.SetPackets(m.getPacketsInOrder())
 		}
 		m.uiState.FilterMode = false
